Crate noirc_errors

Re-exports

pub use reporter::CustomDiagnostic;

pub use reporter::DiagnosticKind;

Modules

call_stack

position 🔒

reporter

Macros

print_to_stderr

Macro to print formatted output to stderr

print_to_stdout

Macro to print formatted output to stdout

println_to_stderr

Macro to print formatted output to stderr

println_to_stdout

Macro to print formatted output to stdout

Structs

Located

Location

Span

Spanned

Functions

print_args_or_exit

print_to_stderr

Print the input to stderr, and exit gracefully if SIGPIPE is received. Rust ignores SIGPIPE by default, converting pipe errors into ErrorKind::BrokenPipe

print_to_stdout

Print the input to stdout, and exit gracefully if SIGPIPE is received. Rust ignores SIGPIPE by default, converting pipe errors into ErrorKind::BrokenPipe

Type Aliases

Position