noirc_errors::call_stack

Struct CallStack

pub struct CallStack(Vec<Location>);

A non-empty list of Locations.

Tuple Fields
0: Vec<Location>
Implementations

impl CallStack

pub fn new(locations: Vec<Location>) -> Self

Construct a new call stack from a potentially empty list of Locations.

pub fn empty() -> Self

Constructor to use when we don’t have location information.

pub fn is_empty(&self) -> bool

Check if the call stack is empty.

A call stack can be non-empty and still end in a dummy location, for example if we are using the SSA parser, with no source files.

pub fn last_or_dummy(&self) -> Location

Get the last location, or a dummy one if it’s empty.

pub fn last(&self) -> Option<&Location>

Get the last location, unless the call stack is empty.

Note that the last location may be a dummy.

pub fn len(&self) -> usize